Refactor vertex-line-angle function
This commit adds a match for the non-vertical case,
and reduces the branches that can panic to two.
Essentially, a case in which both the x values are equal,
and the y values are equal cannot occur, and should be a panic.

The refactor also removes some redundant slope-setting logic and
variables