Add topology-preserving Visvalingam-Whyatt variant
This is based on the technique outlined at
https://www.jasondavies.com/simplify/

It uses an R* tree to restrict the intersection search to only those
segments that fall within a givn triangle's bounding box. If an
intersection is detected, the triangle that causes it is stored for
re-processing, and the next-largest triangle area is used as the
epsilon instead.