Re-factor self-intersection removal logic
Previously, we were retrieving the next-largest triangle, and lowering
its epsilon. This was guaranteed to eventually work; if it was the
wrong point, the intersection would be re-detected, and we'd try
again.

This was, frankly, unacceptable.

The new method introduces a boolean field into the VScore struct. If
an intersection is detected, we update the field. When the
intersection-causing point (b) is
removed, we also update the area of the preceding point (a) to be very
small, ensuring its immediate removal since it will be popped off the
heap next. Removal of (a) causes the (b) triangle to be re-calculated,
removing the self-intersection.