Check for disjoint Polygons without cloning
This adds a method on Polygon which uses the existing get_position
check, omitting interior rings. This is used to represent the Polygon
in simplified form, allowing a candidate point from a Polygon to be
checked for containment. This deals with a corner case in the
Polygon-Polygon distance logic, in which a disjoint candidate Polygon
is contained inside a Polygon's inner ring.